fix: refactor and fix unit tests
