+++ packages/graphback-runtime-knex/src/PgKnexDBDataProvider.ts
@@ -1,8 +1,8 @@
 import { getDatabaseArguments } from '@graphback/core';
 import { GraphQLObjectType } from 'graphql';
 import * as Knex from 'knex';
-import { KnexDBDataProvider } from './KnexDBDataProvider';
 import { NoDataError } from '@graphback/runtime';
+import { KnexDBDataProvider } from './KnexDBDataProvider';
 
 /**
  * Knex.js database data provider exposing basic CRUD operations.
