+++ packages/graphback-runtime/tests/snapshots/LayeredRuntimeResolverCreatorTest.ts.md
@@ -1,0 +1,46 @@
+# Snapshot report for `tests/LayeredRuntimeResolverCreatorTest.ts`
+
+The actual snapshot is saved in `LayeredRuntimeResolverCreatorTest.ts.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## find Todo by text
+
+> Snapshot 1
+
+    {
+      Mutation: {
+        createTodos: Function {},
+        createUser: Function {},
+        deleteTodos: Function {},
+        deleteUser: Function {},
+        updateTodos: Function {},
+        updateUser: Function {},
+      },
+      Query: {
+        findAllTodos: Function {},
+        findAllUsers: Function {},
+        findTodos: Function {},
+        findUsers: Function {},
+      },
+      Subscription: {
+        deletedTodos: {
+          subscribe: Function subscribe {},
+        },
+        deletedUser: {
+          subscribe: Function subscribe {},
+        },
+        newTodos: {
+          subscribe: Function subscribe {},
+        },
+        newUser: {
+          subscribe: Function subscribe {},
+        },
+        updatedTodos: {
+          subscribe: Function subscribe {},
+        },
+        updatedUser: {
+          subscribe: Function subscribe {},
+        },
+      },
+    }
