+++ templates/ts-apollo-fullstack/package.json
@@ -12,12 +12,10 @@
   "license": "MIT",
   "devDependencies": {
     "graphback-cli": "0.11.0-rc1",
-
     "graphql": "14.6.0",
     "tslint": "5.20.1",
     "typescript": "3.7.5"
   },
   "dependencies": {
-    "pg": "7.18.1"
   }
 }
