+++ templates/ts-apollo-fullstack/server/package.json
@@ -9,7 +9,7 @@
     "start": "ts-node src/index.ts"
   },
   "dependencies": {
-    "@graphback/runtime": "^0.11.0-rc1",
+    "@graphback/runtime-knex": "^0.11.0-rc1",
     "@graphql-toolkit/file-loading": "0.9.7",
     "@types/cors": "2.8.6",
     "@types/express": "4.17.2",
